# contracts/openapi/agent-public.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoiceSpeechRequest": {
        "additionalProperties": false,
        "properties": {
          "locale": {
            "default": "en",
            "pattern": "^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$",
            "title": "Locale",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "VoiceSpeechRequest",
        "type": "object"
      },
      "VoiceTranscription": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "pattern": "^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$",
            "title": "Language",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "language"
        ],
        "title": "VoiceTranscription",
        "type": "object"
      },
      "VoiceTranscriptionEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VoiceTranscription"
          },
          "message": {
            "default": "Voice transcription completed.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "SUCCESS",
            "title": "Status",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "VoiceTranscriptionEnvelope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DWP Agent Runtime",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/voice/speech": {
      "post": {
        "operationId": "synthesize_voice_v1_voice_speech_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Roles"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Person-Public-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 36,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Person-Public-Id"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Display-Name-B64",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 400,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Display-Name-B64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceSpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/mpeg": {}
            },
            "description": "Synthesized speech."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Synthesize Voice",
        "tags": [
          "voice"
        ]
      }
    },
    "/v1/voice/transcriptions": {
      "post": {
        "operationId": "transcribe_voice_v1_voice_transcriptions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Content-Type",
            "required": true,
            "schema": {
              "title": "Content-Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Voice-Locale",
            "required": true,
            "schema": {
              "maxLength": 40,
              "minLength": 2,
              "pattern": "^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$",
              "title": "X-Dwp-Voice-Locale",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Permissions",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Permissions"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-User-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Tenant-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Dwp-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Correlation-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Roles",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Roles"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Person-Public-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 36,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Person-Public-Id"
            }
          },
          {
            "in": "header",
            "name": "X-DWP-Display-Name-B64",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 400,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dwp-Display-Name-B64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "contentMediaType": "application/octet-stream",
                "description": "Ephemeral voice recording. The Agent runtime does not persist this body.",
                "maxLength": 4194304,
                "minLength": 1,
                "title": "Audio",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTranscriptionEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcribe Voice",
        "tags": [
          "voice"
        ]
      }
    }
  }
}